Animal issued and expresses genes.
When the "animal" is initialised, it is issued two sets of chromosomes (for ease of typing called genomes) with 10 "genes" each in each chromosome consisting of a random float.
The genes (a tuple of 2 numbers: Gene code and dominance score) are compared and the gene with the highest dominance score is selected, and written to the "active genes" list (this isn't a list of chomosomes, just the genes, I would like to change this in the future.)
The genes are then expressed by changing the size and colour of the animal.

Signed-off-by: Scobbo <robert.coleman@minivoid.net>