Child can now inherit genes.
Pressing space will generate a child based on the two parent animals that are generated on start up.
The simulation will randomly pick which chromosome to copy to the genome being sent to the child from each parent.
The child will then chose which gene from each pair of chromosomes to activate based on it's dominance score.
The child's gene expression will then show one of four possible options for each gene (in this case, the width for example will be one of four outcomes.)

Next I'm going to make the gene expressions make more sense, such as having the genes that determine the colour come from the same chromosome (after more study.)
I'd also like to display a graphical representation of which chromosomes from each parent were passed down, and which genes are active in each animal.

Signed-off-by: Scobbo <robert.coleman@minivoid.net>